+++ test/Framework/Controller/ControllerTest.php
@@ -11,7 +11,7 @@ class ControllerTest extends TestCase
     private $request;
     private $controller;
 
-    public function setUp()
+    public function setUp(): void
     {
         $this->di = $this->createMock('Rauma\Service\Container');
         $this->request = $this->createMock('Psr\Http\Message\ServerRequestInterface');