+++ changelog.md
@@ -7,6 +7,7 @@ Changelog
 * Add `isLoggedIn` method to `Authorisation` class
 * Upgraded all dependencies and moved to PHP 7.1
 * Remove `description` from `Authorisation` in favour of attributes
+* Custom `auth` services can now be configured
 
 3.14
 ----